tout marche SAUF SPLIT PLKHJDFKJS

+++ libft.h
@@ -41,7 +41,7 @@ size_t	ft_strlen(const char *str);
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
 char	*ft_strnstr(const char *big, const char *little, size_t n);
 char	*ft_strrchr(const char *str, int l);
-char	*ft_substr(char const *s, size_t start, size_t len);
+char	*ft_substr(char const *s, unsigned int start, size_t len);
 int		ft_tolower(int l);
 int		ft_toupper(int l);
 char	*ft_strtrim(char const *s1, char const *set);
